Fixed detection of non-available task level 6

+++ app.js
@@ -58,12 +58,17 @@ app.delete("/todos/:id", async function (request, response) {
   // Then, we have to respond back with true/false based on whether the Todo was deleted or not.
   // response.send(true)
   try {
-    await Todo.destroy({
-      where: {
-        id: request.params.id,
-      },
-    });
-    return response.send(true);
+    const todo = await Todo.findByPk(request.params.id);
+    if (todo) {
+      await Todo.destroy({
+        where: {
+          id: request.params.id,
+        },
+      });
+      return response.send(true);
+    } else {
+      return response.send(false);
+    }
   } catch (error) {
     return response.send(false);
   }
